no empty newlines before preprocessor directives
Fixes issue #10
Usually a newline character is inserted at the end of the line before a
preprocessor directive. However this behaviour is counterproductive if that
line is empty.

Lines are represented as elements of an array. Inserting a newline character in
a (otherwise empty) array element would add an unneccesary newline as a final
result.